api_yamdb/requirements.txt
aiosmtplib==2.0.1
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==39.0.1
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
unisender-python-client==0.0.1
urllib3==1.26.14
